Added method to check if the server is rebooting and removed delay
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2295 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/source/http.cpp
@@ -1480,7 +1480,7 @@ int Http::sendHTTPResource(string& uri, int systemrequest, int onlyHeader,
           }
           else
           {
-            int lastSlashOffset = uri.length();
+						u_long lastSlashOffset = uri.length();
             while(lastSlashOffset && uri[lastSlashOffset] != '/')
               --lastSlashOffset;
